+++ test/cli-option-tags/test.mjs
@@ -1,3 +1,5 @@
+import path from 'node:path';
+import assert from 'node:assert/strict';
 import { expect } from '@playwright/test';
 import { test, execPlaywrightTest, TestDir } from '../helpers.mjs';
 
@@ -5,20 +7,25 @@ const testDir = new TestDir(import.meta);
 
 test(testDir.name, () => {
   testDir.clearDir('.features-gen');
-  execPlaywrightTest(
-    testDir.name,
-    'node ../../dist/cli --tags "@include and not @exclude" && npx playwright test',
-  );
-  expect(testDir.isFileExists('.features-gen/skip-no-tags.feature.spec.js')).toEqual(false);
-  expect(testDir.isFileExists('.features-gen/skip-by-top-tag.feature.spec.js')).toEqual(false);
+  execPlaywrightTest(testDir.name, 'node ../../dist/cli --tags "@include and not @exclude"');
 
-  expect(testDir.isFileExists('.features-gen/include.feature.spec.js')).toEqual(true);
-  let fileContents = testDir.getFileContents('.features-gen/include.feature.spec.js');
+  // important to keep included files in separate directory (subdir)
+  // to ensure that directory structure kept the same when running with tags
+
+  testDir.getAllFiles('features').forEach((relativePath) => {
+    const outputFilePath = path.join('.features-gen', `${relativePath}.spec.js`);
+    const fileName = path.basename(relativePath);
+    const shouldExist = !fileName.startsWith('skip');
+    const message = `File: ${outputFilePath} expected to ${shouldExist ? 'exist' : 'not exist'}`;
+    // use assert as playwright's expect does not show custom message for some reason
+    assert(testDir.isFileExists(outputFilePath) === shouldExist, message);
+  });
+
+  let fileContents = testDir.getFileContents('.features-gen/subdir/include.feature.spec.js');
   expect(fileContents).toContain(`test("scenario 1",`);
   expect(fileContents).not.toContain(`test("scenario 2",`);
 
-  expect(testDir.isFileExists('.features-gen/outline.feature.spec.js')).toEqual(true);
-  fileContents = testDir.getFileContents('.features-gen/outline.feature.spec.js');
+  fileContents = testDir.getFileContents('.features-gen/subdir/outline.feature.spec.js');
   expect(fileContents).not.toContain(`test("Example #1",`);
   expect(fileContents).not.toContain(`test("Example #2",`);
   expect(fileContents).toContain(`test("Example #3",`);