+++ test/helpers.mjs
@@ -3,6 +3,8 @@ import assert from 'node:assert/strict';
 import path from 'node:path';
 import test from 'node:test';
 import fs from 'node:fs';
+import fg from 'fast-glob';
+import { fileURLToPath } from 'node:url';
 import { expect } from '@playwright/test';
 
 defineTestOnly(test);
@@ -89,6 +91,11 @@ export class TestDir {
   getFileContents(relativePath) {
     const absPath = this.getAbsPath(relativePath);
     return fs.readFileSync(absPath, 'utf8');
+  }
+
+  getAllFiles(relativePath) {
+    const absPath = fileURLToPath(this.getAbsPath(relativePath));
+    return fg.sync(path.join(absPath, '**')).map((file) => path.relative(absPath, file));
   }
 }